backend_requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pyserial==3.5
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.1
cryptography==41.0.0
